Hooking up pitchfinder pitch detection to AudioAnalyser

+++ src/types/index.ts
@@ -21,7 +21,7 @@ export interface Dimensions {
 export interface AudioData {
   amplitudeAtNote(note: Note): number;
   frequencies: Uint8Array;
-  wave: Uint8Array;
+  wave: Float32Array;
   amplitude: number; // 0 - 1
   hzPerIdx: number;
   notes: NoteInfo[];
